Started building dynamic query page.
"Never tell people how to do things. Tell them what to do and they will
surprise you with their ingenuity.
It is foolish and wrong to mourn the men who died. Rather we should
thank God that such men lived.
Accept the challenges, so that you may feel the exhilaration of
victory."