Added DataTables
Added tables, using UEType as selector on select box. Removed search
button and set to onClick instead. Fading table in and container out.

Working on styling with Laura, datatables are populated.

"The reason it's on the rise is because probably the boom times are
getting even more boomer." - Bertie Ahern