Fixes from demo
Datepickers are all active, so that meets Rachel's level of done for 3
user stories.
Added a progress bar to all imsis between dates query
Goes by the length of the returned data
Have made slight edits to IMSIRestService to accomodate this as I need
to return the length of the json string as a response header in order
for the progress bar to work correctly. It is not fully implemented yet
as I don't have the time or the compute power at home, but the bones are
there and it will be a simple job on Monday to complete. Date validation
is also complete and will be added to all dates queries on Monday. 
Also added a crappy border around the queries that will require some CSS
work, but overall, all suggestions from PO are now dealt with.