Chart everything
All Network Management Engineer queries now have charts that display
relevant data as laid out in US #17. They're not pretty (yet) but it
does fulfill the requirement for the user story. Also wrote manual tests
this morning to accompany them. Secured the charts from being seen by
non-NME user in the event that the query is shared. Will consider with
the team on Monday whether or not we should attempt the drill-down
story.